+++ R/extract.R
@@ -22,10 +22,10 @@ extract <- function(models,
                     gof_map = modelsummary::gof_map,
                     gof_omit = NULL,
                     add_rows = NULL,
-                    add_rows_location = NULL,
                     stars = FALSE,
                     fmt = '%.3f',
                     estimate = 'estimate',
+                    add_rows_location = NULL,
                     ...) {
 
     # models must be a list of models
@@ -40,7 +40,8 @@ extract <- function(models,
     sanity_coef_omit(coef_omit)
     sanity_gof_map(gof_map)
     sanity_gof_omit(gof_omit)
-    sanity_add_rows(add_rows, add_rows_location, models)
+    sanity_add_rows_location(add_rows_location) # deprecated 
+    sanity_add_rows(add_rows, models)
     sanity_stars(stars)
     sanity_fmt(fmt)
     sanity_estimate(estimate)
@@ -132,31 +133,32 @@ extract <- function(models,
     gof <- gof[order(idx, gof$term),] 
 
     # add_rows: this needs to be done after sorting and combining to preserve
-    # user-selected row order
+    # user-selected row order. We need the 'section' information otherwise we
+    # won't know where to draw the line which separates estimates from gof. 
     if (!is.null(add_rows)) {
-        add_rows <- lapply(add_rows, as.character) %>%  # TODO: remove once sanity checks are complete
-                    do.call('rbind', .) %>%
-                    data.frame(stringsAsFactors = FALSE) %>%
-                    stats::setNames(c('term', model_names)) %>%
-                    dplyr::mutate(group = 'gof', statistic = '')
-        add_rows <- add_rows[, colnames(gof)]
 
-        # sanity check add_rows_location
-        checkmate::assert_numeric(add_rows_location, null.ok = TRUE,
-                                  max.len = 1, lower = 0, upper = nrow(gof))
-        if (is.null(add_rows_location)) { # bottom if location is not specified
-            gof <- dplyr::bind_rows(gof, add_rows)
-        } else {
-            if (add_rows_location == 0) { # top if location is 0
-                gof <- dplyr::bind_rows(add_rows, gof)
-            } else if (add_rows_location == nrow(gof)) { # bottom if location is nrow(gof)
-                gof <- dplyr::bind_rows(gof, add_rows)
-            } else { # middle otherwise
-                top <- gof[1:add_rows_location,]
-                bot <- gof[(add_rows_location + 1):nrow(gof),]
-                gof <- dplyr::bind_rows(top, add_rows, bot)
+        if (inherits(add_rows, 'list')) {
+            add_rows <- do.call('rbind', add_rows) %>%
+                        data.frame(stringsAsFactors = FALSE) %>%
+                        stats::setNames(colnames(gof)[2:ncol(gof)]) %>%
+                        dplyr::mutate(section = 'bottom', position = 1e6)
+        }
+
+        for (i in 1:nrow(add_rows)) {
+            newrow <- add_rows[i, ] %>% dplyr::select(-section, -position)
+            if (add_rows$section[i] == 'bottom') {
+                newrow <- newrow %>% dplyr::mutate(group = 'gof') 
+                gof <- gof %>%
+                       tibble::add_row(newrow, .before = add_rows$position[i])
+            } else {
+                newrow <- newrow %>% 
+                          dplyr::mutate(group = 'estimates',
+                                 statistic = 'estimate') 
+                est <- est %>%
+                       tibble::add_row(newrow, .before = add_rows$position[i])
             }
         }
+
     }
 
     # combine estimates and gof